requirements.txt
numpy 
pandas 
setuptools
mysql-connector-python
scikit-learn
lightgbm
xgboost
tensorboard #for model monitoring and comparison
torch
matplotlib
mlflow # for experiment tracking
dvc
flask